+++ src/methods.jl
@@ -55,43 +55,63 @@ function is_same(A::AbstractString, B::AbstractString)
 end
 
 """
+    OIFITS.is_same(A, B, sym)
+
+yields whether field `sym` of data-blocks `A` and `B` are sufficiently
+identical for OI-FITS data.
+
+"""
+@inline is_same(A::T, B::T, sym::Symbol) where {T<:OIDataBlock} =
+    is_same(A, B, Val(sym))
+
+is_same(A::T, B::T, ::Val{S}) where {T<:OIDataBlock,S} = begin
+    if (defined = isdefined(A, S)) != isdefined(B, S)
+        return false
+    elseif defined
+        return is_same(getfield(A, S), getfield(B, S))
+    else
+        return true
+    end
+end
+
+"""
     OIFITS.fix_name(str)
 
 yields string `str` converted to uppercase letters and with trailing spaces
-removed.  This is to follows FITS conventions that names comparisons should be
-done ignoring of the case of letters and trailing spaces.
+removed.  This is to follow FITS conventions that names comparisons should be
+done ignoring the case of letters and trailing spaces.
 
 """
 fix_name(str::AbstractString) = uppercase(rstrip(str))
 
-# Extend A[name] syntax to get an OI-ARRAY, OI-WAVELENGTH, or OI-CORR by its
+# Extend A[name] syntax to get an OI_ARRAY, OI_WAVELENGTH, or OI_CORR by its
 # name (according to FITS conventions).  The syntax keys(String,A) can be used
 # to get an iterator over the string keys of A.
-for (T, field) in ((OI_ARRAY,      :arrname),
-                   (OI_WAVELENGTH, :insname),
-                   (OI_CORR,       :corrname))
+for (type, name) in ((:OI_ARRAY,      :arrname),
+                     (:OI_WAVELENGTH, :insname),
+                     (:OI_CORR,       :corrname))
     @eval begin
-        getindex(A::AbstractArray{<:$T}, key::AbstractString) = begin
+        getindex(A::AbstractArray{<:$type}, key::AbstractString) = begin
             @inbounds for i in eachindex(A)
-                if is_same(A[i].$field, key)
+                if is_same(A[i].$name, key)
                     return A[i]
                 end
             end
             throw(KeyError(fix_name(key)))
         end
-        keys(::Type{String}, A::AbstractArray{<:$T}) =
-            Iterators.map(x -> fix_name(x.$field), A)
-        haskey(A::AbstractArray{<:$T}, key::AbstractString) = begin
+        keys(::Type{String}, A::AbstractArray{<:$type}) =
+            Iterators.map(x -> fix_name(x.$name), A)
+        haskey(A::AbstractArray{<:$type}, key::AbstractString) = begin
             @inbounds for i in eachindex(A)
-                if is_same(A[i].$field, key)
+                if is_same(A[i].$name, key)
                     return true
                 end
             end
             return false
         end
-        get(A::AbstractArray{<:$T}, key::AbstractString, def) = begin
+        get(A::AbstractArray{<:$type}, key::AbstractString, def) = begin
             @inbounds for i in eachindex(A)
-                if is_same(A[i].$field, key)
+                if is_same(A[i].$name, key)
                     return A[i]
                 end
             end
@@ -115,64 +135,115 @@ type of the result:
 
 """
 extname(db::OIDataBlock) = extname(typeof(db))
-extname(T::Type{<:Union{String,Symbol}}, db::OIDataBlock) =
-    extname(T, typeof(db))
+extname(T::Type{<:OIDataBlock}) = extname(String, T)
+extname(S::Type{<:Union{String,Symbol}}, db::OIDataBlock) =
+    extname(S, typeof(db))
 
-for (type, ext) in ((:OI_TARGET,     :OI_TARGET),
-                    (:OI_ARRAY,      :OI_ARRAY),
-                    (:OI_WAVELENGTH, :OI_WAVELENGTH),
-                    (:OI_CORR,       :OI_CORR),
-                    (:OI_VIS,        :OI_VIS),
-                    (:OI_VIS2,       :OI_VIS2),
-                    (:OI_T3,         :OI_T3),
-                    (:OI_FLUX,       :OI_FLUX),
-                    (:OI_INSPOL,     :OI_INSPOL))
+for sym in (:OI_TARGET,
+            :OI_ARRAY,
+            :OI_WAVELENGTH,
+            :OI_CORR,
+            :OI_VIS,
+            :OI_VIS2,
+            :OI_T3,
+            :OI_FLUX,
+            :OI_INSPOL)
     @eval begin
-        extname(::Type{<:$type}) = $(String(ext))
-        extname(::Type{String}, ::Type{<:$type}) = $(String(ext))
-        extname(::Type{Symbol}, ::Type{<:$type}) = $(QuoteNode(ext))
+        extname(::Type{String}, ::Type{<:$sym}) = $(String(sym))
+        extname(::Type{Symbol}, ::Type{<:$sym}) = $(QuoteNode(sym))
     end
 end
 
+"""
+    OIFITS.get_list(T, ds)
+
+yields the list (a regular vector) of entries of type `T` stored by the OI-FITS
+data-set `ds`.
+
+"""
+get_list(::Type{OITargetEntry}, ds::OIDataSet) = get_list(ds.target)
+get_list(::Type{OI_ARRAY},      ds::OIDataSet) = ds.array
+get_list(::Type{OI_WAVELENGTH}, ds::OIDataSet) = ds.instr
+get_list(::Type{OI_CORR},       ds::OIDataSet) = ds.correl
+get_list(::Type{OI_VIS},        ds::OIDataSet) = ds.vis
+get_list(::Type{OI_VIS2},       ds::OIDataSet) = ds.vis2
+get_list(::Type{OI_T3},         ds::OIDataSet) = ds.t3
+get_list(::Type{OI_FLUX},       ds::OIDataSet) = ds.flux
+get_list(::Type{OI_INSPOL},     ds::OIDataSet) = ds.inspol
+
+"""
+    OIFITS.get_dict(T, ds)
+
+yields the dictionary associated to entries of type `T` stored by the OI-FITS
+data-set `ds`.
+
+"""
+get_dict(::Type{OITargetEntry}, ds::OIDataSet) = ds.target_dict
+get_dict(::Type{OI_ARRAY},      ds::OIDataSet) = ds.array_dict
+get_dict(::Type{OI_WAVELENGTH}, ds::OIDataSet) = ds.instr_dict
+get_dict(::Type{OI_CORR},       ds::OIDataSet) = ds.correl_dict
+
 #------------------------------------------------------------------------------
 # PROPERTIES
+#
+# We use the `db.key` syntax to give access to other properties (e.g.,
+# `extname`, `name`, `eff_wave`, or `eff_band`) than the fields of the
+# data-block `db`.
+#
+# To allow for efficient handling of the `db.key` syntax, we wrap the `key`
+# symbol as a `Val(key)` so that each `getproperty` or `setproperty!` method
+# can be specific to the value of `key`.
 
-propertynames(db::OIDataBlock) = (fieldnames(typeof(db))..., :extname)
+propertynames(db::OIDataBlock) =
+    (fieldnames(typeof(db))..., :extname)
+
+propertynames(db::Union{OI_ARRAY,OI_WAVELENGTH,OI_CORR}) =
+    (fieldnames(typeof(db))..., :extname, :name)
+
+propertynames(db::Union{OI_VIS,OI_VIS2,OI_T3,OI_FLUX,OI_INSPOL}) =
+    (fieldnames(typeof(db))..., :extname, :eff_wave, :eff_band)
 
 getproperty(db::OIDataBlock, sym::Symbol) = getproperty(db, Val(sym))
 getproperty(db::OIDataBlock, ::Val{S}) where {S} = getfield(db, S)
 getproperty(db::OIDataBlock, ::Val{:extname}) = extname(typeof(db))
 
-# Indirections to instrument (OI_WAVELENGTH) for OI_VIS, OI_VIS2, OI_T3, and
-# OI_FLUX.
-for type in (:OI_VIS, :OI_VIS2, :OI_T3, :OI_FLUX)
-    @eval begin
-        @eval propertynames(db::$type) =
-            (fieldnames(typeof(db))..., :extname, :eff_wave, :eff_band)
-        @eval getproperty(db::$type, ::Val{:eff_wave}) =
-            getfield(db, :instr).eff_wave
-        @eval getproperty(db::$type, ::Val{:eff_band}) =
-            getfield(db, :instr).eff_band
+setproperty!(db::OIDataBlock, sym::Symbol, val) =
+    setproperty!(db, Val(sym), val)
+
+setproperty!(db::T, ::Val{S}, val) where {S,T<:OIDataBlock} = begin
+    try
+        setfield!(db, S, convert(fieldtype(T, S), val))
+    catch ex
+        rethrow_convert_field_error(ex, T, sym, val)
     end
 end
 
-setproperty!(x::OIDataBlock, f::Symbol, v) = _set_field!(x, f, v)
-
-@inline _get_field(A, sym::Symbol, def=undef) =
-    isdefined(A, sym) ? getfield(A, sym) : def
-
-@inline _convert_field(obj::Any, sym::Symbol, val) =
-    _convert_field(typeof(obj), sym, val)
-
-@inline _convert_field(T::Type, sym::Symbol, val) =
-    try
-        return convert(fieldtype(T, sym), val)
-    catch ex
-        _rethrow_convert_field_error(ex, T, sym, val)
+for (type, name) in ((:OI_ARRAY,      :arrname),
+                     (:OI_WAVELENGTH, :insname),
+                     (:OI_CORR,       :corrname))
+    @eval begin
+        getproperty(db::$type, ::Val{:name}) = db.$name
+        setproperty!(db::$type, ::Val{:name}, val) = (db.$name = val)
     end
+end
 
-@noinline function _rethrow_convert_field_error(ex::Exception, T::Type,
-                                                sym::Symbol, val)
+for type in (:OI_VIS, :OI_VIS2, :OI_T3, :OI_FLUX)
+    @eval begin
+        # FIXME: Directly calling `getfield` is necessary to optimize the
+        #        indirection (speedup: 1.4ns instead of 370ns).
+        getproperty(db::$type, ::Val{:eff_wave}) =
+            getfield(db, :instr).eff_wave
+        getproperty(db::$type, ::Val{:eff_band}) =
+            getfield(db, :instr).eff_band
+        setproperty!(db::$type, ::Val{:eff_wave}, val) =
+            getfield(db, :instr).eff_wave = val
+        setproperty!(db::$type, ::Val{:eff_band}, val) =
+            getfield(db, :instr).eff_band = val
+    end
+end
+
+@noinline function rethrow_convert_field_error(ex::Exception, T::Type,
+                                               sym::Symbol, val)
     if isa(ex, MethodError) && ex.f === convert
         rethrow(ArgumentError(string(
             "Cannot `convert` an object of type `", typeof(val),
@@ -183,37 +254,8 @@ setproperty!(x::OIDataBlock, f::Symbol, v) = _set_field!(x, f, v)
     end
 end
 
-# `_set_field!(obj,sym,val)` is similar to `setfield!(obj,sym,val)` except that
-# conversion of value `val` is automatically done and that error message is
-# more informative if conversion is not possible.
-@inline _set_field!(obj, sym::Symbol, val) =
-    setfield!(obj, sym, _convert_field(obj, sym, val))
-
-@inline function _define_field!(dst, src, sym::Symbol, val=undef)
-    if val !== undef
-         _set_field!(dst, sym, val)
-    elseif isdefined(src, sym)
-         _set_field!(dst, sym, getfield(src, sym))
-    end
-    nothing
-end
-
 #------------------------------------------------------------------------------
-# BUILDING OF DATA-SETS
-
-function OIDataSet(args::OIDataBlock...)
-    # Create empty instance, then push the dependecy-less data-blocks and,
-    # finally, the data-blocks with dependencies.
-    data = OIDataSet(undef)
-    for flag in (false, true)
-        for db in args
-            if isa(db, DataBlocksWithDependencies) == flag
-                push!(data, db)
-            end
-        end
-    end
-    return data
-end
+# COPY DATA-BLOCKS AND DATA-SETS
 
 # Copy the outer structure of an OI-FITS data-block.
 function copy(A::T) where {T<:OIDataBlock}
@@ -226,122 +268,322 @@ function copy(A::T) where {T<:OIDataBlock}
     return B
 end
 
-"""
-    push!(data::OIDataSet, args::OIDataBlock...) -> data
+function copy(A::OIDataSet)
+    # Create new instance and copy contents.  Including the dictionary for
+    # re-writing target identifiers.  The copy is thus in the same "state" as
+    # the original.
+    B = OIDataSet()
+    _copy!(B.target,        A.target)
+    _copy!(B.target_dict,   A.target_dict)
+    _copy!(B.target_id_map, A.target_id_map)
+    _copy!(B.array,         A.array)
+    _copy!(B.array_dict,    A.array_dict)
+    _copy!(B.instr,         A.instr)
+    _copy!(B.instr_dict,    A.instr_dict)
+    _copy!(B.correl,        A.correl)
+    _copy!(B.correl_dict,   A.correl_dict)
+    _copy!(B.vis,           A.vis)
+    _copy!(B.vis2,          A.vis2)
+    _copy!(B.t3,            A.t3)
+    _copy!(B.flux,          A.flux)
+    _copy!(B.inspol,        A.inspol)
+    return B
+end
 
-pushes OI-FITS data-blocks `args...` in `data`.  This method ensures that
-`data` (and its contents) remain consistent.  If a data-block has undefined
-dependencies, they must be already part of `data`.  If a data-block has defined
-dependencies, they will be merged with those already stored in `data`.
-
-"""
-function push!(data::OIDataSet, args::OIDataBlock...)
-    for db in args
-        push!(data, db)
+function _copy!(dest::Vector{T}, src::Vector{T}) where {T}
+    if dest !== src && length(src) > 0
+        copyto!(resize!(dest, length(src)), src)
     end
-    return data
+    return dest
 end
 
-# FIXME:
-function push!(data::OIDataSet, db::OI_TARGET)
-    isdefined(data, :target) && error("OI_TARGET already defined")
-    setfield!(data, :target, db)
-    return data
+function _copy!(dest::Dict{K,V}, src::Dict{K,V}) where {K,V}
+    empty!(dest)
+    for (key, val) in src
+        dest[key] = val
+    end
+    return dest
 end
 
-# Extend push! for OI_ARRAY, OI_WAVELENGTH, and OI_CORR.
-for (type, name, field) in (
-    (:OI_ARRAY,      QuoteNode(:arrname),  QuoteNode(:array)),
-    (:OI_WAVELENGTH, QuoteNode(:insname),  QuoteNode(:instr)),
-    (:OI_CORR,       QuoteNode(:corrname), QuoteNode(:correl)))
-    @eval begin
-        function push!(data::OIDataSet, db::$type)
-            isdefined(db, $name) || throw_undefined_field($name)
-            other = get(getfield(data, $field), getfield(db, $name), nothing)
-            if other === nothing
-                push!(getfield(data, $field), db)
-            elseif other != db # FIXME:
-                error(string("attempt to replace already existing ",
-                             extname($type), " named \"", name, "\""))
-            end
-            return data
+function _copy!(dest::OI_TARGET, src::OI_TARGET)
+    dest.revn = src.revn
+    _copy!(get_list(dest), get_list(src))
+    return dest
+end
+
+#------------------------------------------------------------------------------
+# PUSH IN DATA-SETS
+
+# Create empty instance, then push all data-blocks which must be ordered.
+OIDataSet(args::OIDataBlock...) = push!(OIDataSet(), args...)
+
+for (type, pass) in ((:OI_TARGET,     1),
+                     (:OI_ARRAY,      1),
+                     (:OI_WAVELENGTH, 1),
+                     (:OI_CORR,       1),
+                     (:OI_VIS,        2),
+                     (:OI_VIS2,       2),
+                     (:OI_T3,         2),
+                     (:OI_FLUX,       2),
+                     (:OI_INSPOL,     2))
+    @eval function _push!(ds::OIDataSet, db::$type, pass::Integer)
+        if pass == $pass
+            push!(ds, db)
         end
+        nothing
     end
 end
 
-for (type, field) in ((:OI_VIS,    :vis),
-                      (:OI_VIS2,   :vis2),
-                      (:OI_T3,     :t3),
-                      (:OI_FLUX,   :flux),
-                      (:OI_INSPOL, :inspol))
-    @eval begin
-        function push!(data::OIDataSet, db::$type)
-            # Datablock must be copied on write to avoid side-effects.
-            copy_on_write = true
+"""
+    push!(ds::OIDataSet, db) -> ds
 
-            # Set/fix OI_WAVELENGTH dependency.
-            isdefined(db, :insname) || throw_undefined_field(:insname)
-            db, copy_on_write = _set_dependency!(
-                data.instr, db, copy_on_write, Val(:insname), Val(:instr))
+adds an OI-FITS data-block `db` in the data-set `ds` and returns `ds`.  After
+the operation, the contents of `db` is left unchanged but may be partially
+shared by `ds`.  Depending on type of `db`, different things can happen:
 
-            # Set/fix OI_ARRAY dependency.
-            if isdefined(db, :arrname)
-                db, copy_on_write = _set_dependency!(
-                    data.array, db, copy_on_write, Val(:arrname), Val(:array))
-            end
+- If `db` is an `OI_TARGET` instance, the targets from `db` not already in the
+  data-set `ds` are added to the list of targets in `ds` and the internal
+  dictionary in `ds` mapping target identifiers in `db` to those in `ds` is
+  reinitialized for subsequent data-blocks.
 
-            # Set/fix OI_CORREL dependency.
-            if $type !== OI_INSPOL && isdefined(db, :corrname)
-                db, copy_on_write = _set_dependency!(
-                    data.correl, db, copy_on_write, Val(:corrname), Val(:correl))
-            end
+- If `db` is an `OI_ARRAY`, `OI_WAVELENGTH`, or `OI_CORR` instance, it is added
+  to the corresponding list of data-blocks in the data-set `ds` unless an entry
+  with a name matching that of `db` already exists in `ds`.  In this latter
+  case, nothing is done except checking that the two data-blocks with matching
+  names have the same contents.  This is need to ensure the consistency of the
+  data-set.
 
-            push!(data.$field, db)
-            return data
-        end
-    end
-end
+- If `db` is an `OI_VIS`, `OI_VIS2`, `OI_T3`, `OI_FLUX`, or `OI_INSPOL`
+  instance, it is added to the corresponding list of data-blocks in the
+  data-set `ds` after having rewritten its target identifiers according to the
+  mapping set by the last `OI_TARGET` pushed into the data-set `ds`.  Add
+  keyword `rewrite_target_id=false` to avoid rewritting target identifiers.
 
-@noinline throw_undefined_field(name::Union{AbstractString,Symbol}) =
-    throw(ErrorException("undefined field `$name`"))
-
-function _set_dependency!(deps::AbstractVector, # list of dependencies known by data-set
-                          db::T,                # datablock to be pushed on data-set
-                          copy_on_write::Bool,  # copy datablock on write?
-                          ::Val{name},          # field name of dependency name
-                          ::Val{value}          # field name of dependency value
-                          ) where {T<:OIDataBlock,name,value}
-
-    other = get(deps, getfield(db, name), nothing)
-    if other === nothing
-        # Dependency not yet in list.
-        if isdefined(db, value)
-            push!(deps, getfield(db, value))
+"""
+function push!(ds::OIDataSet, db::OI_TARGET)
+    # Add all new targets found in `db` to those defined in `ds` and build a
+    # dictionary to reindex target identifiers in `db` and subsequent
+    # data-blocks.
+    name_set      = Set{String}()    # to check unicity of names in source
+    id_set        = Set{Int}()       # to check unicity of identifiers in source
+    target_id_map = ds.target_id_map # to rewrite target_id in source
+    target_list   = ds.target.list   # list of unique targets
+    target_dict   = ds.target_dict   # mapping of target names to identifiers
+    length(target_list) == length(target_dict) || error(
+        "inconsistent list and dictionary of targets")
+    empty!(target_id_map) # reset target identifier mapping
+    for tgt in get_list(db)
+        name = fix_name(tgt.target)
+        name ∈ name_set && error(
+            "duplicate target name \"", name, "\" in ", db.extname)
+        push!(name_set, name)
+        id = Int(tgt.target_id)
+        id ∈ id_set && error(
+            "duplicate target identifier ", id, " in ", db.extname)
+        push!(id_set, id)
+        if haskey(target_dict, name)
+            # The same target already existed in destination.
+            new_id = target_dict[name]
         else
-            error("no ", extname(eltype(deps)), " named \"",
-                  fix_name(getfield(db, name)), "\" found")
+            # This target is new.
+            new_id = length(target_list) + 1
+            push!(target_list, OITargetEntry(tgt; target_id = new_id))
+            target_dict[name] = new_id
         end
+        target_id_map[id] = new_id
+    end
+    # Fix revision number.
+    ds.target.revn = max(ds.target.revn, db.revn)
+    return ds
+end
+
+function push!(ds::OIDataSet,
+               db::T) where {T<:Union{OI_ARRAY,OI_WAVELENGTH,OI_CORR}}
+    dict = get_dict(T, ds)
+    list = get_list(T, ds)
+    name = fix_name(db.name)
+    if haskey(dict, name)
+        # Make sure the two data-blocks with the same name are
+        # identical.
+        assert_identical(db, list[dict[name]])
     else
-        # Another dependency with same name already in data-set.  Check that the
-        # two versions are identical.
-        setfield = true
-        if isdefined(db, value)
-            if getfield(db, value) === other
-                sefield = false
-            elseif getfield(db, value) == other
-                error("multiple ", extname(eltype(deps)), " named \"",
-                      fix_name(getfield(db, name)), "\" found")
-            end
-        end
-        if setfield
-            if copy_on_write
-                db = copy(db)
-                copy_on_write = false
-            end
-            setfield!(db, value, other)
+        # Push new data-block in destination.
+        dict[name] = length(push!(list, db))
+    end
+    return ds
+end
+
+function push!(ds::OIDataSet, db::T;
+               rewrite_target_id::Bool = true) where {
+                   T<:Union{OI_VIS,OI_VIS2,OI_T3,OI_FLUX,OI_INSPOL}}
+    new_db = copy(db) # make a copy to avoid side effects
+    if rewrite_target_id
+        new_db.target_id = rewrite_indices(db.target_id, ds.target_id_map)
+    else
+        id_min, id_max = extrema(db.target_id)
+        if id_min < 1 || id_max > length(db.target)
+            error("out of range target identifier(s) (", id_min, ":", id_max,
+                  " ⊈ 1:", length(db.target), ")")
         end
     end
-    return db, copy_on_write
+    if isdefined(new_db, :arrname)
+        new_db.array = _find_depencency(
+            ds.array, ds.array_dict, new_db.arrname)
+    end
+    if isdefined(new_db, :insname)
+        new_db.instr = _find_depencency(
+            ds.instr, ds.instr_dict, new_db.insname)
+    end
+    if isdefined(new_db, :corrname)
+        new_db.correl = _find_depencency(
+            ds.correl, ds.correl_dict, new_db.corrname)
+    end
+    push!(get_list(T, ds), new_db)
+    return ds
+end
+
+function _find_depencency(dest_list::Vector{T},
+                          dest_dict::Dict{String,Int},
+                          dep::AbstractString) where {S,T<:OIDataBlock}
+    name = fix_name(dep)
+    haskey(dest_dict, name) || error(
+        "no extensions ", extname(T), " match name \"", name, "\"")
+    return dest_list[dest_dict[name]]
+end
+
+"""
+    OIFITS.rewrite_indices(inds, dict) -> inds′
+
+yields array of indices similar to `inds` and rewritten according to the
+dictionary `dict` used as a mapping from integers to integers.  If the indices
+are left unchanged, the input array is returned.
+
+"""
+function rewrite_indices(inds::AbstractArray{<:Integer,N},
+                         dict::Dict{<:Integer,<:Integer}) where {N}
+    new_inds = similar(inds)
+    unchanged = true
+    for i in eachindex(new_inds, inds)
+        idx = inds[i]
+        new_idx = dict[idx]
+        new_inds[i] = new_idx
+        unchanged &= (new_idx == idx)
+    end
+    return (unchanged ? inds : new_inds)
+end
+
+"""
+    OIFITS.assert_identical(A, B)
+
+throws an exception if *named* data-blocks `A` and `B` are not identical.  A
+*named* data-block is an `OI_ARRAY`, `OI_WAVELENGTH`, or `OI_CORR` extension
+that can be uniquely identifed by its name.
+
+The variant
+
+    OIFITS.assert_identical(A, B, sym)
+
+is to assert that fields `A.sym` and `B.sym` are identical, it shall only be
+called if `A` and `B` have the same type and the same name.
+
+""" assert_identical
+
+function assert_identical(A::T, B::T) where {T<:Union{OI_ARRAY,OI_WAVELENGTH,OI_CORR}}
+    if !(A === B)
+        is_same(A.name, B.name) || throw_assertion_error(
+            "two ", extname(T), " data-blocks named \"", A.name,
+            "\" and \"", B.name, "\" cannot be identical")
+        assert_identical(A, B, *)
+    end
+end
+
+function assert_identical(A::OI_ARRAY, B::OI_ARRAY, ::typeof(*))
+    assert_identical(A, B, :revn)
+    assert_identical(A, B, :frame)
+    assert_identical(A, B, :arrayx)
+    assert_identical(A, B, :arrayy)
+    assert_identical(A, B, :arrayz)
+    assert_identical(A, B, :tel_name)
+    assert_identical(A, B, :sta_name)
+    assert_identical(A, B, :sta_index)
+    assert_identical(A, B, :diameter)
+    assert_identical(A, B, :staxyz)
+    assert_identical(A, B, :fov)
+    assert_identical(A, B, :fovtype)
+end
+
+function assert_identical(A::OI_WAVELENGTH, B::OI_WAVELENGTH, ::typeof(*))
+    assert_identical(A, B, :revn)
+    assert_identical(A, B, :eff_wave)
+    assert_identical(A, B, :eff_band)
+end
+
+function assert_identical(A::OI_CORR, B::OI_CORR, ::typeof(*))
+    assert_identical(A, B, :revn)
+    assert_identical(A, B, :ndata)
+    assert_identical(A, B, :iindx)
+    assert_identical(A, B, :jindx)
+    assert_identical(A, B, :corr)
+end
+
+@inline assert_identical(A::T, B::T, sym::Symbol) where {T<:NamedDataBlock} =
+    assert_identical(A, B, Val(sym))
+
+assert_identical(A::T, B::T, val::Val{sym}) where {T<:NamedDataBlock,sym} =
+    is_same(A, B, val) || throw_assertion_error(
+        "two ", extname(T), " extensions both named \"", fix_name(A.name),
+        " have different field `", sym, "`")
+
+throw_assertion_error(msg::AbstractString) = throw(AssertionError(msg))
+@noinline throw_assertion_error(args...) =
+    throw_assertion_error(string(args...))
+
+#------------------------------------------------------------------------------
+# MERGE DATA-SETS
+
+merge(A::OIDataSet) = copy(A)
+merge(A::OIDataSet, args::OIDataSet...) = merge!(OIDataSet(), A, args...)
+
+merge!(A::OIDataSet) = A
+merge!(A::OIDataSet, others::OIDataSet...) = begin
+    for B in others
+        merge!(A, B)
+    end
+    return A
+end
+
+function merge!(A::OIDataSet, B::OIDataSet)
+    # First push targets and dependencies.
+    push!(A, B.target)
+    for db in B.array
+        push!(A, db)
+    end
+    for db in B.instr
+        push!(A, db)
+    end
+    for db in B.correl
+        push!(A, db)
+    end
+
+    # Then push all other data-blocks.
+    for db in B.vis
+        push!(A, db)
+    end
+    for db in B.vis2
+        push!(A, db)
+    end
+    for db in B.t3
+        push!(A, db)
+    end
+    for db in B.flux
+        push!(A, db)
+    end
+    for db in B.inspol
+        push!(A, db)
+    end
+
+    # Return updated data-set.
+    return A
 end
 
 #------------------------------------------------------------------------------
@@ -350,10 +592,10 @@ end
 """
     OITargetEntry([def::OITargetEntry]; kwds...)
 
-yields an entry (a row) of the OI_TARGET table in an OI-FITS file.  All fields
-are specified by keywords.  If template entry `def` is specified, keywords have
+yields an entry of the `OI_TARGET` table in an OI-FITS file.  All fields are
+specified by keywords.  If template entry `def` is specified, keywords have
 default values taken from `def`; otherwise all keywords are mandatory but
-`category` which is assmed to be `""` if unspecified.
+`category` which is assumed to be `""` if unspecified.
 
 """
 function OITargetEntry(;
@@ -434,25 +676,63 @@ function OITargetEntry(def::OITargetEntry;
                          category)
 end
 
-rows(db::OI_TARGET) = getfield(db, :rows)
+"""
+    OI_TARGET(lst=OITargetEntry[]; revn=0)
+
+yields an `OI_TARGET` data-block.  Optional argument `lst` is a vector of
+`OITargetEntry` specifying the targets (none by default).  Keyword `revn`
+specifies the revision number.
+
+An instance, say `db`, of `OI_TARGET` has the following properties:
+
+    db.revn   # revision number
+    db.list   # list of targets
+
+and can be used as an iterable or as an array to access the targets
+individually by their index or by their name:
+
+    length(db) # the number of targets
+    db[i]      # the i-th target
+    db[key]    # the target whose name matches string `key`
+
+    for tgt in db; ...; end # loop over all targets
+
+Note that target identifiers (field `target_id`) are automatically rewritten to
+be identical to the index in the target list.
+
+Standard methods `get` and `haskey` work as aexpected and according to the type
+(integer or string) of the key.  For the `keys` method, the default is to
+return an iterator over the target names, but the type of the expected keys
+can be specified:
+
+    keys(db)          # iterator over target names
+    keys(String, db)  # idem
+    keys(Integer, db) # iterator over target indices
+    keys(Int, db)     # idem
+
+Call [`OIFITS.get_column`](@ref) to retrieve a given target field for all
+targets of `OI_TARGET` data-block in the form of a vector.
+
+""" OI_TARGET
+
+get_list(db::OI_TARGET) = db.list
 
 # Make OI_TARGET instances iterable and behave more or less like vectors
 # (with properties).
-
-ndims(db::OI_TARGET) = ndims(rows(db))
-size(db::OI_TARGET) = size(rows(db))
-size(db::OI_TARGET, i) = size(rows(db), i)
-axes(db::OI_TARGET) = axes(rows(db))
-axes(db::OI_TARGET, i) = axes(rows(db), i)
-length(db::OI_TARGET) = length(rows(db))
+ndims(db::OI_TARGET)   =  ndims(get_list(db))
+size(db::OI_TARGET)    =   size(get_list(db))
+size(db::OI_TARGET, i) =   size(get_list(db), i)
+axes(db::OI_TARGET)    =   axes(get_list(db))
+axes(db::OI_TARGET, i) =   axes(get_list(db), i)
+length(db::OI_TARGET)  = length(get_list(db))
 
 IndexStyle(db::OI_TARGET) = IndexStyle(OI_TARGET)
-IndexStyle(::Type{OI_TARGET}) = IndexStyle(fieldtype(OI_TARGET, :rows))
+IndexStyle(::Type{OI_TARGET}) = IndexStyle(fieldtype(OI_TARGET, :list))
 
 eltype(db::OI_TARGET) = eltype(OI_TARGET)
 eltype(::Type{OI_TARGET}) = OITargetEntry
 
-@inline @propagate_inbounds getindex(db::OI_TARGET, i::Integer) = rows(db)[i]
+@inline @propagate_inbounds getindex(db::OI_TARGET, i::Integer) = get_list(db)[i]
 
 function getindex(db::OI_TARGET, name::AbstractString)
     for tgt in db
@@ -464,9 +744,9 @@ function getindex(db::OI_TARGET, name::AbstractString)
           db.extname, "data-block")
 end
 
-firstindex(db::OI_TARGET) = firstindex(rows(db))
-lastindex(db::OI_TARGET) = lastindex(rows(db))
-eachindex(db::OI_TARGET) = eachindex(rows(db))
+firstindex(db::OI_TARGET) = firstindex(get_list(db))
+lastindex(db::OI_TARGET) = lastindex(get_list(db))
+eachindex(db::OI_TARGET) = eachindex(get_list(db))
 
 function iterate(db::OI_TARGET,
                  state::Tuple{Int,Int} = (firstindex(db),
@@ -475,7 +755,7 @@ function iterate(db::OI_TARGET,
     if i > n
         return nothing
     else
-        return db.rows[i], (i + 1, n)
+        return get_list(db)[i], (i + 1, n)
     end
 end
 
@@ -512,6 +792,7 @@ values(db::OI_TARGET) = db
 keys(db::OI_TARGET) = Iterators.map(x -> fix_name(x.target), db)
 keys(::Type{String}, db::OI_TARGET) = keys(db)
 keys(::Type{Int}, db::OI_TARGET) = eachindex(db)
+keys(::Type{Integer}, db::OI_TARGET) = eachindex(db)
 
 """
     OIFITS.get_column([T,] db::OI_TARGET, col)